rag_local/phase2-retrieval/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sentence-transformers==3.0.0
psycopg2-binary==2.9.9
elasticsearch==8.13.0
weaviate-client==4.6.4
httpx==0.27.0